Refactor into embedded driver
