+++ pkg/ocm/share/repository/json/json.go
@@ -624,6 +624,6 @@ func (m *mgr) UpdateReceivedShare(ctx context.Context, user *userpb.User, share 
 	return rs, nil
 }
 
-func (m *mgr) ProcessEmbeddedShare(ctx context.Context, user *userpb.User, share *ocm.ReceivedShare) (*ocm.ReceivedShare, error) {
-	return nil, errtypes.NotSupported("processing embedded shares is not supported")
+func (m *mgr) GetEmbeddedPayload(ctx context.Context, user *userpb.User, share *ocm.ReceivedShare) (string, error) {
+	return "", errtypes.NotSupported("operation not supported")
 }
