+++ pkg/ocm/share/repository/nextcloud/nextcloud.go
@@ -483,6 +483,6 @@ func (sm *Manager) do(ctx context.Context, a Action, username string) (int, []by
 	return resp.StatusCode, body, nil
 }
 
-func (sm *Manager) ProcessEmbeddedShare(ctx context.Context, user *userpb.User, share *ocm.ReceivedShare) (*ocm.ReceivedShare, error) {
-	return nil, errtypes.NotSupported("processing embedded shares is not supported")
+func (sm *Manager) GetEmbeddedPayload(ctx context.Context, user *userpb.User, share *ocm.ReceivedShare) (string, error) {
+	return "", errtypes.NotSupported("operation not supported")
 }
